Addressed comment about creating new files for artifacts,executions and lineage
